add AnalysisHelper
